PMM-6077 move types
