+++ pmm-app/src/plugin.json
@@ -22,7 +22,7 @@
   },
   "includes": [
     { "type": "dashboard", "name": "pmm-qan", "path": "dashboards/pmm-qan.json" },
-    { "type": "panel", "name": "pmm-qan" },
+    { "type": "panel", "name": "pmm-qan-app-panel" },
     { "type": "dashboard", "name": "PMM Inventory", "path": "dashboards/pmm-inventory.json" },
     { "type": "panel", "name": "pmm-inventory-panel" },
     { "type": "dashboard", "name": "PMM Settings", "path": "dashboards/pmm-settings.json" },
