+++ pmm-app/src/pmm-qan/plugin.json
@@ -1,5 +1,5 @@
 {
   "type": "panel",
   "name": "QAN",
-  "id": "pmm-qan"
+  "id": "pmm-qan-app-panel"
 }
